services/gateway/requirements.txt
fastapi[all]
pydantic
pyjwt
